Keep track of search steps
Also add a property test that asserts that dpll always uses the same or fewer search steps than backtracking. What's not surprising is that this is always true; what _is_ surprising is _how much better_ it is - dpll often reaches conclusions with *zero* search steps (which I've equated to speculative assignments).